refactor npmignore checks for consistency
Using as similar logic as possible is the next best thing to using the same code.

+++ bin/slt-release.sh
@@ -115,7 +115,7 @@ then
   git checkout "$TAG"
   # npm uses .gitignore if there is no .npmignore, so we'll use that as
   # our starting point if there isn't already a .npmigore file
-  if [[ ! -f ".npmignore" ]]; then
+  if test -f ".gitignore" -a ! -f ".npmignore"; then
     cp .gitignore .npmignore
   fi
   # ignore the entire test tree
